refactor(ui): move featurePost components into post component
